pyproject.toml
[build-system]
requires = [ "setuptools", "wheel",]
build-backend = "setuptools.build_meta"

[project]
name = "python-devops"
requires-python = ">=3.7"
version = "0.3.1"
description = "Material for a python devops course."
dependencies = [ "black==24.2.0", "blinker==1.7.0", "build==1.0.3", "click==8.1.7", "coverage==7.4.1", "Flask==3.0.2", "iniconfig==2.0.0", "itsdangerous==2.1.2", "Jinja2==3.1.3", "MarkupSafe==2.1.5", "mypy-extensions==1.0.0", "packaging==23.2", "pathspec==0.12.1", "platformdirs==4.2.0", "pluggy==1.4.0", "pyproject_hooks==1.0.0", "pytest==8.0.0", "python-dotenv==1.0.1", "toml==0.10.2", "Werkzeug==3.0.1",]
[[project.authors]]
name = "Sckathach"

